% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabnb_example_counts}
\alias{cabnb_example_counts}
\title{CAB-NB example count matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 40 rows and 30 columns.
}
\source{
Simulated example data.
}
\usage{
data(cabnb_example_counts)
}
\description{
An example microbial count matrix for demonstrating the CAB-NB workflow.
Rows represent samples and columns represent taxa (a 40 by 30 integer matrix).
}
\keyword{datasets}
