% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabnb_example_metadata}
\alias{cabnb_example_metadata}
\title{CAB-NB example sample metadata}
\format{
An object of class \code{data.frame} with 40 rows and 2 columns.
}
\source{
Simulated example data.
}
\usage{
data(cabnb_example_metadata)
}
\description{
Sample-level metadata corresponding to \code{cabnb_example_counts}
(a data frame with 40 rows and 2 columns.).
}
\keyword{datasets}
